+++ src/misc/TreeView/TreeItem.vue
@@ -37,21 +37,22 @@
         console.log('Item clicked. this.data.name = ')
         console.log(this.model.name)
 
+        var wasSelected = this.model.selected
         this.$emit('unselectAll')
-        this.model.selected = true
+        this.model.selected = !wasSelected
         console.log('this.data =')
         console.log(this.model)
         var category = []
         category.unshift(this.model.name)
-        this.$emit('clicked', category)
+        this.$emit('clicked', { category: category, isSelected: this.model.selected })
       },
-      handleChildClicked (category) {
-        console.log('handleChildClicked() called. category = ')
-        console.log(category)
+      handleChildClicked (payload) {
+        console.log('handleChildClicked() called. payload = ')
+        console.log(payload)
 
-        category.unshift(this.model.name)
+        payload.category.unshift(this.model.name)
         // Bubble event upwards (should stop at TreeView component)
-        this.$emit('clicked', category)
+        this.$emit('clicked', payload)
       },
       handleUnselectAll () {
         this.$emit('unselectAll')