+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added ability to clear category selection by clicking on selected category for a second time, closes #163.
+
 ### Fixed
 - Fixed bug where 'Two Coordinate Geodesics' calculator had old URL of '.../calc/distance-between-two-coordinates', closes #164.
 - Changed justified text in calculator selection grid description to left-aligned, closes #161.