+++ changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed bug where 'Two Coordinate Geodesics' calculator had old URL of '.../calc/distance-between-two-coordinates', closes #164.
 - Changed justified text in calculator selection grid description to left-aligned, closes #161.
 - Changed cursor when over category tree view to 'clickable' (e.g. index finger), closes #165.
+- Made output variables more easily recognizable as outputs, closes #166.
 
 ## [v3.2.1] - 2018-02-27